+++ src/Examples/example.py
@@ -1,0 +1,1 @@
+import pmpy.simulation.discrete_event as des