+++ src/pmpy/example.py
@@ -1,3 +1,0 @@
-import pmpy.simulation.discrete_event as des
-import pmpy.simulation.monte_carlo as mc
-import pmpy.simulation.monte_carlo as a