+++ src/pmpy/simulation/__init__.py
@@ -1,3 +1,3 @@
 """Simulation subpackage includes "discrete_event", "distributions" and "monte_carlo" modules.
 """
-print("Simulation subpackage includes `discrete_event`, `distributions` and `monte_carlo` modules.")
+print("Simulation subpackage includes discrete_event, distributions and monte_carlo modules.")